requirements.txt
biopython
more-itertools
numpy
pandas
pybedtools
scipy
tqdm
openpyxl